+++ examples/Console/STEM/AI/AI_continuous.cs
@@ -1,7 +1,6 @@
 /// AI_continuous.cs with synchronous mode.
 ///
-/// This example demonstrates the process of obtaining AI data in continuous mode.
-/// Additionally, it utilizes a loop to retrieve AI data with 8 channels from STEM.
+/// This example demonstrates the process of obtaining AI data in continuous mode with 8 channels from STEM.
 ///
 /// To begin with, it demonstrates the steps to open the AI and configure the AI parameters.
 /// Next, it outlines the procedure for reading the streaming AI data.
@@ -32,26 +31,6 @@ using WPC.Product;
 
 class STEM_AI_continuous
 {
-    static void loop_func(STEM handle, int slot, int samples, int delay, int exit_time)
-    {
-        List<List<double>> streaming_list;
-        int time_cal = 0;
-        while (time_cal < exit_time)
-        {
-            // Read data acquisition
-            streaming_list = handle.AI_readStreaming(slot, samples, delay);
-
-            foreach (List<double> sample in streaming_list)
-            {
-                Console.WriteLine(string.Format("[{0}]", string.Join(", ", sample)));
-            }
-
-            // Wait
-            Thread.Sleep(delay); // delay [ms]
-            time_cal += delay;
-        }
-    }
-
     static public void Main()
     {
         // Get C# driver version
@@ -79,8 +58,11 @@ class STEM_AI_continuous
             int err;
             int slot = 1; // Connect AIO module to slot
             int mode = Const.AI_MODE_CONTINUOUS;
-            float sampling_rate = 1000;
+            float sampling_rate = 200;
+            int read_points = 200;
+            int delay = 200;   // ms
             int timeout = 3000; // ms
+            List<int> chip_select = new List<int>() {0, 1};
 
             // Get firmware model & version
             string[] driver_info = dev.Sys_getDriverInfo(timeout:timeout);
@@ -106,35 +88,37 @@ class STEM_AI_continuous
             Console.WriteLine($"AI_open in slot {slot}: {err}");
 
             // Enable CS
-            err = dev.AI_enableCS(slot, new List<int> {0, 1}, timeout:timeout);
+            err = dev.AI_enableCS(slot, chip_select, timeout:timeout);
             Console.WriteLine($"AI_enableCS in slot {slot}: {err}");
 
-            // Set AI acquisition mode to continuous
+            // Set AI acquisition mode to continuous mode
             err = dev.AI_setMode(slot, mode, timeout:timeout);
             Console.WriteLine($"AI_setMode {mode} in slot {slot}: {err}");
 
-            // Set AI sampling rate to 1k (Hz)
+            // Set AI sampling rate
             err = dev.AI_setSamplingRate(slot, sampling_rate, timeout:timeout);
             Console.WriteLine($"AI_setSamplingRate {sampling_rate} in slot {slot}: {err}");
 
-            // Start AI acquisition
+            // Start AI
             err = dev.AI_start(slot, timeout:timeout);
             Console.WriteLine($"AI_start in slot {slot}: {err}");
 
-            // Wait for data
-            Thread.Sleep(1000);
-
-            // loop parameters
-            int get_samples = 600;
-            int delay = 50;
-            int exit_time = 300;
-
-            // Start loop
-            loop_func(dev, slot, get_samples, delay, exit_time);
+            // Wait a while for data acquisition
+            Thread.Sleep(1000); // delay [ms]
 
             // Stop AI
             err = dev.AI_stop(slot, timeout:timeout);
             Console.WriteLine($"AI_stop in slot {slot}: {err}");
+
+            int data_len = 1;
+            while (data_len > 0){
+                // Read data acquisition
+                List<List<double>> ai_2Dlist = dev.AI_readStreaming(slot, read_points, delay:delay);
+                Console.WriteLine($"number of samples = {ai_2Dlist.Count}");
+
+                // Update data len
+                data_len = ai_2Dlist.Count;
+            }
 
             // Close AI
             err = dev.AI_close(slot, timeout:timeout);