+++ examples/Console/STEM/AIO/AIO_one_channel_loopback.cs
@@ -61,6 +61,7 @@ class STEM_AIO_one_channel_loopback
             int err;
             int slot = 1; // Connect AIO module to slot
             int timeout = 3000; // ms
+            List<double> ao_value_list = new List<double>() {0, 0.5, 1, 1.5, 2, 2.5, 3, 3.5};
 
             // Get firmware model & version
             string[] driver_info = dev.Sys_getDriverInfo(timeout:timeout);
@@ -85,41 +86,41 @@ class STEM_AIO_one_channel_loopback
             err = dev.AI_open(slot, timeout:timeout);
             Console.WriteLine($"AI_open in slot {slot}: {err}");
 
-            // Open AO
-            err = dev.AO_open(slot, timeout:timeout);
-            Console.WriteLine($"AO_open in slot {slot}: {err}");
-
             // Enable CS
             err = dev.AI_enableCS(slot, new List<int> {0, 1}, timeout:timeout);
             Console.WriteLine($"AI_enableCS in slot {slot}: {err}");
 
-            // Read data acquisition
-            List<double> sample = dev.AI_readOnDemand(slot, timeout:timeout);
-
-            // Print data
-            Console.WriteLine(string.Format("[{0}]", string.Join(", ", sample)));
-
-            // Write AO vaule 1.5(V) in channel 0
-            err = dev.AO_writeOneChannel(slot, 0, 1.5, timeout:timeout);
-            Console.WriteLine($"AO_writeOneChannel in ch0 in slot {slot}: {err}");
-
-            // Write AO vaule 2.5(V) in channel 1
-            err = dev.AO_writeOneChannel(slot, 1, 2.5, timeout:timeout);
-            Console.WriteLine($"AO_writeOneChannel in ch1 in slot {slot}: {err}");
-
-            // Write AO vaule 3.5(V) in channel 2
-            err = dev.AO_writeOneChannel(slot, 2, 3.5, timeout:timeout);
-            Console.WriteLine($"AO_writeOneChannel in ch2 in slot {slot}: {err}");
-
-            // Write AO vaule 4.5(V) in channel 3
-            err = dev.AO_writeOneChannel(slot, 3, 4.5, timeout:timeout);
-            Console.WriteLine($"AO_writeOneChannel in ch3 in slot {slot}: {err}");
+            // Open AO
+            err = dev.AO_open(slot, timeout:timeout);
+            Console.WriteLine($"AO_open in slot {slot}: {err}");
 
             // Read data acquisition
-            sample = dev.AI_readOnDemand(slot, timeout:timeout);
+            List<double> ai_list = dev.AI_readOnDemand(slot, timeout:timeout);
 
             // Print data
-            Console.WriteLine(string.Format("[{0}]", string.Join(", ", sample)));
+            Console.WriteLine(string.Format("[{0}]", string.Join(", ", ai_list)));
+
+            // Write AO vaule in channel 0
+            err = dev.AO_writeOneChannel(slot, 0, ao_value_list[0], timeout:timeout);
+            Console.WriteLine($"In slot {slot} channel 0, the AO value is {ao_value_list[0]}: {err}");
+
+            // Write AO vaule in channel 1
+            err = dev.AO_writeOneChannel(slot, 1, ao_value_list[1], timeout:timeout);
+            Console.WriteLine($"In slot {slot} channel 1, the AO value is {ao_value_list[1]}: {err}");
+
+            // Write AO vaule in channel 2
+            err = dev.AO_writeOneChannel(slot, 2, ao_value_list[2], timeout:timeout);
+            Console.WriteLine($"In slot {slot} channel 2, the AO value is {ao_value_list[2]}: {err}");
+
+            // Write AO vaule in channel 3
+            err = dev.AO_writeOneChannel(slot, 3, ao_value_list[3], timeout:timeout);
+            Console.WriteLine($"In slot {slot} channel 3, the AO value is {ao_value_list[3]}: {err}");
+
+            // Read data acquisition
+            ai_list = dev.AI_readOnDemand(slot, timeout:timeout);
+
+            // Print data
+            Console.WriteLine(string.Format("[{0}]", string.Join(", ", ai_list)));
 
             // Close AI
             err = dev.AI_close(slot, timeout:timeout);