+++ examples/Console/STEM/AO/AO_write_one_channel.cs
@@ -45,6 +45,7 @@ class STEM_AO_write_one_channel
             int err;
             int slot = 1; // Connect AIO module to slot
             int timeout = 3000; // ms
+            List<double> ao_value_list = new List<double>() {0, 0.5, 1, 1.5, 2, 2.5, 3, 3.5};
 
             // Get firmware model & version
             string[] driver_info = dev.Sys_getDriverInfo(timeout:timeout);
@@ -69,21 +70,21 @@ class STEM_AO_write_one_channel
             err = dev.AO_open(slot, timeout:timeout);
             Console.WriteLine($"AO_open in slot {slot}: {err}");
 
-            // Write AO 1.5(V) in channel 0
-            err = dev.AO_writeOneChannel(slot, 0, 1.5, timeout:timeout);
-            Console.WriteLine($"AO_writeOneChannel in ch0 in slot {slot}: {err}");
+            // Write AO vaule in channel 0
+            err = dev.AO_writeOneChannel(slot, 0, ao_value_list[0], timeout:timeout);
+            Console.WriteLine($"In slot {slot} channel 0, the AO value is {ao_value_list[0]}: {err}");
 
-            // Write AO 2.5(V) in channel 1
-            err = dev.AO_writeOneChannel(slot, 1, 2.5, timeout:timeout);
-            Console.WriteLine($"AO_writeOneChannel in ch1 in slot {slot}: {err}");
+            // Write AO vaule in channel 1
+            err = dev.AO_writeOneChannel(slot, 1, ao_value_list[1], timeout:timeout);
+            Console.WriteLine($"In slot {slot} channel 1, the AO value is {ao_value_list[1]}: {err}");
 
-            // Write AO 3.5(V) in channel 2
-            err = dev.AO_writeOneChannel(slot, 2, 3.5, timeout:timeout);
-            Console.WriteLine($"AO_writeOneChannel in ch2 in slot {slot}: {err}");
+            // Write AO vaule in channel 2
+            err = dev.AO_writeOneChannel(slot, 2, ao_value_list[2], timeout:timeout);
+            Console.WriteLine($"In slot {slot} channel 2, the AO value is {ao_value_list[2]}: {err}");
 
-            // Write AO 4.5(V) in channel 3
-            err = dev.AO_writeOneChannel(slot, 3, 4.5, timeout:timeout);
-            Console.WriteLine($"AO_writeOneChannel in ch3 in slot {slot}: {err}");
+            // Write AO vaule in channel 3
+            err = dev.AO_writeOneChannel(slot, 3, ao_value_list[3], timeout:timeout);
+            Console.WriteLine($"In slot {slot} channel 3, the AO value is {ao_value_list[3]}: {err}");
 
             // Close AO
             err = dev.AO_close(slot, timeout:timeout);