+++ examples/Console/USBDAQF1AOD/AIO/AO_output_while_AI_streaming.cs
@@ -1,0 +1,3 @@
+
+// USBDAQF1AOD did not support this example code
+