+++ examples/Console/WifiDAQF4A/AI/AI_on_demand_once.cs
@@ -61,10 +61,10 @@ class WifiDAQF4A_AI_on_demand_once
             Console.WriteLine($"AI_setMode {mode}: {err}");
 
             // Read data acquisition acquisition
-            List<double> sample = dev.AI_readOnDemand(port, timeout:timeout);
+            List<double> ai_list = dev.AI_readOnDemand(port, timeout:timeout);
 
             // Print data
-            Console.WriteLine(string.Format("[{0}]", string.Join(", ", sample)));
+            Console.WriteLine(string.Format("[{0}]", string.Join(", ", ai_list)));
 
             // Close AI
             err = dev.AI_close(port, timeout:timeout);